% Generated by roxygen2 (4.1.0.9001): do not edit by hand
\name{contrib_cache}
\alias{contrib_cache}
\alias{contrib_cache_clear}
\title{Repository Indexes}
\usage{
contrib_cache(repos, type = getOption("pkgType"))

contrib_cache_clear(...)
}
\arguments{
  \item{repos}{Repository URL}

  \item{type}{Package type (see \code{\link[utils]{contrib.url}})}

  \item{...}{parameters passed to \code{contrib_cache}.}
}
\description{
  \code{contrib_cache} returns the path to the RDS file were a repository index is
stored when first accessed by \code{\link[utils]{available.packages}}.

\code{contrib_cache_clear} deletes the index cache file,
so that the repository's index is refreshed by the next call to
\code{\link[utils]{available.packages}}.
}

