% Generated by roxygen2 (4.1.1): do not edit by hand
\name{download_file}
\alias{download_file}
\alias{with_rcurl}
\title{Downloading Files From Password Protected Directories}
\usage{
download_file(url, destfile, ...)

with_rcurl(expr)
}
\arguments{
  \item{url}{A character string naming the URL of a resource to be
    downloaded.}

  \item{destfile}{A character string with the name where the downloaded
    file is saved.  Tilde-expansion is performed.}

  \item{...}{arguments passed to \code{\link{download.file}}.}

  \item{expr}{an expression}
}
\description{
  Downloads files using a custom \code{curl} binary that uses
\pkg{RCurl} and is able to download files from password protected
directories.

\code{with_rcurl} executes an expression with the custom \code{curl} binary
setup as the default download method.
}

