% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches-devtools.R
\name{get_remote_package_description}
\alias{get_remote_package_description}
\title{Fetch DESCRIPTION File from Remote}
\usage{
get_remote_package_description(remote, url, path, user = remote$auth_token,
  password = NULL, ...)
}
\arguments{
\item{remote}{remote S3 object as returned by \code{devtools:::remote}}

\item{url}{base repo host url}

\item{path}{path to DESCRIPTION file}

\item{user}{username to use for authentication}

\item{password}{authentication password or token}

\item{...}{arguments passed to \link[httr:GET]{httr::GET}}
}
\description{
Fetch DESCRIPTION File from Remote
}
