% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{get_shim_parent}
\alias{get_shim_parent}
\title{Retrieving Original Patched Function}
\usage{
get_shim_parent(name, envir = NULL)
}
\arguments{
\item{name}{name of the patched function.
Can be specified as \code{package_name::funname}, in which case
argument \code{envir} default value becomes "package_name".}

\item{envir}{environment where the function should be looked for.}
}
\description{
Returns the original function that has been patched.
}
