% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{github_remote_description}
\alias{github_remote_description}
\title{Fetching Github Package Description}
\usage{
github_remote_description(..., remote = NULL,
  url = "https://raw.githubusercontent.com")
}
\arguments{
\item{...}{arguments passed to `devtools:::github_remote` to build a `remote` S3 object
if argument `remote=NULL`.}

\item{remote}{remote S3 object as returned by `devtools:::github_remote`.}

\item{url}{base url}
}
\value{
an S3 object as returned by [devtools::as.package].
}
\description{
Fetches DESCRIPTION file form a remote Github repository.
}
