% Generated by roxygen2 (4.1.1): do not edit by hand
\name{GRAN.update_drat}
\alias{GRAN.update_drat}
\title{Updates Repository Indexes of Built Packages}
\usage{
GRAN.update_drat(outdir = ".", type = c("all", "source", "mac.binary",
  "win.binary"), update = FALSE, repos = NULL)
}
\arguments{
  \item{outdir}{path to output directory}
}
\description{
  
}
\details{
  Create a package repository that aggregates all GitHub source and
\code{drat} repositories into a single one, compatible with
\code{\link{install.packages}}.

GRAN repositories are served by \url{http://gran.r-forge.r-project.org},
which provides access to:
\itemize{
\item user-specific repositories, which favors packages distributed
by a given GitHub user, e.g. \url{http://gran.r-forge.r-project.org/renozao};
\item a top-level cross-user repository, which only includes packages distributed by
their owner (forked repositories), \url{http://gran.r-forge.r-project.org};
\item development/source repository, \url{http://gran.r-forge.r-project.org/devel}
}


The top-level repository
}

