% Generated by roxygen2 (4.1.0.9001): do not edit by hand
\name{gran_dependencies}
\alias{gran_dependencies}
\title{Computes GRAN Package Dependencies}
\usage{
gran_dependencies(pkgs, repos = getOption("repos"), ..., verbose = FALSE)
}
\arguments{
  \item{pkgs}{A character vector of package keys, e.g., \code{'renozao/repotools'} or
\code{'renozao/NMF/devel'}.}

  \item{repos}{CRAN-like repositories urls}

  \item{...}{other parameters passed to \code{\link[utils]{available.packages}}.}

  \item{verbose}{logical that toggles log messages.}
}
\description{
  Computes GRAN Package Dependencies
}
\examples{
\dontshow{# roxygen generated flag
options(R_CHECK_RUNNING_EXAMPLES_=TRUE)
}

gran_dependencies(c('renozao/NMF/devel', 'renozao/repotools'))
}

