% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{OS_type}
\alias{OS_type}
\title{Determinate Type of Operating System}
\usage{
OS_type()
}
\value{
a single character string: code{'unix'}, code{'mac'}, or code{'windows'}
depending on the OS of the calling machine.
}
\description{
Returns the type of OS.
}
\examples{
\dontshow{
options(R_CHECK_RUNNING_EXAMPLES_=TRUE) ## roxygen generated flag
}

## OS_type

OS_type()


}
