% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{parse_deps_versions}
\alias{parse_deps_versions}
\title{Parse Package Dependencies}
\usage{
parse_deps_versions(string, full = FALSE, all = FALSE)
}
\arguments{
\item{string}{Specification of a set of dependencies either as a character vector
or as a single string as returned by \link[devtools:as.package]{devtools::as.package} in fields \code{Depends},
\code{Imports} or \code{Suggests}.}

\item{full}{logical that indicates if version requirements should also be returned.}

\item{all}{loigical that indicates if R requirement and base packages should be
indluded in the result.}
}
\value{
a character vector if \code{full=FALSE} or a character matrix otherwise.
}
\description{
Parse Package Dependencies
}
\examples{
\dontshow{
options(R_CHECK_RUNNING_EXAMPLES_=TRUE) ## roxygen generated flag
}


parse_deps_versions("a, b, c")
# with some version requirement
parse_deps_versions("a (>= 1.2.3), b, c (<= 2.5)")

}
