% Generated by roxygen2 (4.1.1): do not edit by hand
\name{r_versions_n}
\alias{r_versions_n}
\title{Fetch the Last N R Versions}
\usage{
r_versions_n(n = Inf, digits = 2L, cache = TRUE)
}
\arguments{
  \item{n}{Number of versions to return}

  \item{digits}{Number of version digits: 1 = major version number,
2 = <major>.<minor1>, 3 = <major>.<minor1>.<minor2>}

  \item{cache}{logical that indicates if the result should be retrieved from
cache or (re-)fetched from the SVN server.}
}
\description{
  This function uses \code[rversions]{r_versions} to fetch all the
R versions from R main SVN server.
}

