% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-auth.R
\name{remote_get_api}
\alias{remote_get_api}
\title{API Query to Remote Source Control Repositories}
\usage{
remote_get_api(remote, url, path, user = remote$auth_user,
  password = remote$auth_token, ..., config = NULL, error = FALSE)
}
\arguments{
\item{remote}{a repository specification or \code{remote} object.}

\item{url}{base API url to use.}

\item{path}{path to the file to downlad, relatively to the root of the repository.}

\item{user, password}{repository authetication credentials. The default is to use the user define in the remote
specification \code{'auth_user'} and \code{'auth_token'} respectively, which can be determined transparently
with \link{remote_auth}, via the credentials stored in a \code{.netrc} file.}

\item{...}{other arguments passed to \link[httr:GET]{httr::GET}}

\item{error}{logical that indicates if an error should be thrown if the request is unsuccessful}
}
\value{
the request result object returned by \link[httr:GET]{httr::GET}
}
\description{
API Query to Remote Source Control Repositories
}
