% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-auth.R
\name{remote_get_raw}
\alias{remote_get_raw}
\title{Fetching Raw Contents form Remote Source Control Repositories}
\usage{
remote_get_raw(remote, path, url = NULL, ..., destfile = NULL)
}
\arguments{
\item{remote}{a repository specification or \code{remote} object.}

\item{path}{path to the file to downlad, relatively to the root of the repository.}

\item{url}{base url to use. When \code{NULL} (default), the base url is taken from the remote
specification \code{'host_raw'}.}

\item{...}{other arguments passed to \link[httr:GET]{httr::GET}}

\item{destfile}{file where to save the fetched content.
If \code{NULL} (default), then the file is saved in the temporary directory, with a unique prefix of the form
\code{<prefix>__<basename(path)>}.}

\item{user, password}{repository authetication credentials. The default is to use the user define in the remote
specification \code{'auth_user'} and \code{'auth_token'} respectively, which can be determined transparently
with \link{remote_auth}, via the credentials stored in a \code{.netrc} file.}
}
\value{
the full normalized path to the file where the raw content was saved.
}
\description{
Fetching Raw Contents form Remote Source Control Repositories
}
