% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{repos_auth}
\alias{repos_auth}
\title{CRAN-style Repository Authentication Credentials}
\usage{
repos_auth(..., save = FALSE, cache = NULL, exact = FALSE,
  .netrc = NULL)
}
\arguments{
\item{...}{repository keys or credential specifications}

\item{save}{indicates if the new credentials are to be saved on disk,
by updating in the \code{.netrc} file, or in a cache -- which expires at the end of 
the \pkg{R} session.
It can also be a character string, which then specifies the path where to save the 
result (cached entries are not written to disk).

Use \code{save=""} to print to the console.}

\item{cache}{specifies how the temporary cached credentials should be 
loaded: 

\itemize{
\item \code{NULL} loads both the cache and netrc data, with cache override matching 
netrc credentials;
\item \code{TRUE} loads only the cache data; 
\item \code{FALSE} loads only the netrc data.
}

The special value \code{cache = NA} will delete the cache, prior to any other operation.}

\item{exact}{logical that indicates if the the repository keys should be matched 
exactly. If \code{FALSE}, then keys are matched independently of possible trailing 
path specification.}

\item{.netrc}{path or connection to a \emph{netrc} file, which is looked up for
matching repository/host entries.
Default is to look for file \code{'.netrc'} in the user's home directory.

Repository and host credential entries are defined in an \emph{netrc} file format (see \code{\link{read_netrc}} for more details).

Repository shortcut keys start a \code{'@'}, e.g., \code{'@myRepo/path/to/repo'} and are substituted by the URL
\code{'http://<login>:<password>@<machine><path>/path/to/repo'}, where \emph{<machine>}, \emph{<path>},
\emph{<login>} and \code{<password>} are taken from the matched \emph{netrc} specification.

Regular hosts are also matched against the \emph{netrc} specification to add credentials if necessary.}
}
\description{
Gets/sets authentication credentials for CRAN-style.
}
\seealso{
Other auth: \code{\link{read_netrc}},
  \code{\link{repos_url}}
}
