% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{sourceURL}
\alias{sourceURL}
\title{Sourcing Remote Files}
\usage{
sourceURL(url)
}
\arguments{
\item{url}{URL}
}
\description{
This function plays well with CNTLM proxies, because it download the complete file,
before sourcing it locally.
}
