% Generated by roxygen2 (4.1.0.9001): do not edit by hand
\name{GRAN.update_github}
\alias{GRAN.update_github}
\title{Updates GitHub Source Package Repository}
\usage{
GRAN.update_github(src, force = FALSE, fields = GRAN.fields(),
  update = FALSE, actions = c("changes", "PACKAGES", "index"),
  clean = TRUE, test = FALSE, verbose = TRUE)
}
\arguments{
  \item{src}{path to the directory that holds the package directory tree}
}
\description{
  Updates GitHub Source Package Repository
}

