% Generated by roxygen2 (4.1.0.9001): do not edit by hand
\name{sourceURL}
\alias{sourceURL}
\title{Sourcing Remote Files}
\usage{
sourceURL(url)
}
\arguments{
  \item{url}{URL}
}
\description{
  This function plays well with CNTLM proxies, because it download the complete file,
before sourcing it locally.
}

